Split acceptance tests
